refactor: move helper functions and styles to own file
