+++ libs/frontend/application/element/src/css-editor/tailwind-class-editor/TailwindClassEditor.tsx
@@ -1,20 +1,15 @@
 import { CloseOutlined } from '@ant-design/icons'
-import type { IElement } from '@codelab/frontend/abstract/core'
+import type { IElementModel } from '@codelab/frontend/abstract/domain'
 import { Select, Space } from 'antd'
 import type { CustomTagProps } from 'rc-select/lib/BaseSelect'
 import React, { useState } from 'react'
-import styled from 'styled-components'
+import * as Styled from './styles'
 import {
+  extractVariant,
   getOnlyColorValue,
   getValidTailwindClasses,
   isColorClass,
 } from './tailwind-utils'
-
-const Label = styled.span`
-  display: inline-flex;
-  align-items: center;
-  height: 32px;
-`
 
 const getInitialTailwindClasses = () => {
   return getValidTailwindClasses('').map((className) => ({
@@ -23,7 +18,11 @@ const getInitialTailwindClasses = () => {
   }))
 }
 
-export const TailwindClassEditor = ({ element }: { element: IElement }) => {
+export const TailwindClassEditor = ({
+  element,
+}: {
+  element: IElementModel
+}) => {
   const [options, setOptions] = useState<
     Array<{ value: string; label: string }>
   >(getInitialTailwindClasses())
@@ -71,7 +70,7 @@ export const TailwindClassEditor = ({ element }: { element: IElement }) => {
 
   return (
     <>
-      <Label>Tailwind Classes :</Label>
+      <Styled.Label>Tailwind Classes :</Styled.Label>
       <Select
         allowClear
         defaultValue={element.tailwindClassNames ?? []}
@@ -107,46 +106,17 @@ export const TailwindClassEditor = ({ element }: { element: IElement }) => {
   )
 }
 
-const CustomOption = styled.div`
-  padding: 0px 6px;
-  background-color: #edebeb;
-  border: 1px solid #d9d8d8;
-  border-radius: 12px;
-  margin: 2px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  gap: 2px;
-
-  span {
-    font-size: 12px;
-  }
-`
-
 const TagRender = (props: CustomTagProps) => {
   const { onClose, value } = props
 
   return (
-    <CustomOption>
+    <Styled.CustomOption>
       <ColorBox color={value} withPlaceholder={false} />
       {value}
       <CloseOutlined onClick={onClose} />
-    </CustomOption>
+    </Styled.CustomOption>
   )
 }
-
-const StyledColorBox = styled.div`
-  width: 10px;
-  height: 10px;
-  border: 0.3px solid lightgray;
-`
-
-const StyledBlankColorBox = styled.div`
-  width: 10px;
-  height: 10px;
-  background-color: transparent;
-`
 
 const ColorBox = ({
   color,
@@ -156,35 +126,10 @@ const ColorBox = ({
   withPlaceholder?: boolean
 }) => {
   if (!isColorClass(color)) {
-    return withPlaceholder ? <StyledBlankColorBox /> : null
+    return withPlaceholder ? <Styled.BlankColorBox /> : null
   }
 
   const backgroundClassName = `bg-${getOnlyColorValue(color)}`
 
-  return <StyledColorBox className={backgroundClassName} />
-}
-
-interface VariantResult {
-  className: string
-  variant: string
-}
-
-const extractVariant = (input: string): VariantResult => {
-  // Trim the input string
-  const trimmedInput = input.trim()
-  // Match the pattern and extract the required portion
-  // Modified regex to capture up to the last colon
-  const match = trimmedInput.match(/^(.*?):(?=[^:]+$)/)
-
-  if (match) {
-    return {
-      className: trimmedInput.slice(match[0].length),
-      variant: match[1] + ':',
-    }
-  }
-
-  return {
-    className: trimmedInput,
-    variant: '',
-  }
+  return <Styled.ColorBox className={backgroundClassName} />
 }
